components/second-brain/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "second-brain"
version = "0.1.0"
description = "Personal knowledge base maintained by Claude, graph-backed, markdown-as-truth."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Jay" }]
dependencies = [
  "click>=8.1",
  "pydantic>=2.8",
  "ruamel.yaml>=0.18",
  "duckdb>=1.1.0",
  "opendataloader-pdf>=2.2.0",
  "httpx>=0.27",
  "anthropic>=0.40",
  "python-slugify>=8.0",
  "readability-lxml>=0.8",
  "watchdog>=4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5",
  "ruff>=0.6",
  "mypy>=1.11",
]
mlx = [
  "mlx-lm>=0.31,<0.32; platform_system == 'Darwin' and platform_machine == 'arm64'",
]
vectors = [
  "sqlite-vec>=0.1",
  "sentence-transformers>=2.0",
]

[project.scripts]
sb = "second_brain.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/second_brain"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
asyncio_mode = "auto"
markers = [
  "integration: slower end-to-end tests",
]

[tool.coverage.run]
branch = true
source = ["src/second_brain"]

[tool.coverage.report]
show_missing = true
fail_under = 75
